requirements.txt
yt-dlp
moviepy
youtube-transcript-api
Pillow
tqdm
